requirements.txt
aiobreaker==1.2.0
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
boto3==1.24.66
botocore==1.27.66
cachetools==5.3.1
case-converter==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
decorator==5.1.1
Django==4.2.5
django-appconf==1.0.*
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
google-auth==2.23.2
googleapis-common-protos==1.56.2
grpcio==1.59.0
grpcio-reflection==1.47.5
grpcio-tools==1.47.5
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.16.1
itemadapter==0.8.0
itemloaders==1.1.0
jaeger-client==4.8.0
jdatetime==4.1.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
kubernetes==18.20.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mock==5.1.0
mypy-protobuf==3.2.0
oauthlib==3.2.2
opentracing==2.4.0
packaging==23.2
parsel==1.8.1
parso==0.8.3
persian-tools==0.0.11
persiantools==3.0.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==3.20.3
psycopg==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybreaker==0.6.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
s3transfer==0.6.2
Scrapy==2.11.0
sentry-sdk==1.32.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
threadloop==1.0.2
thrift==0.16.0
tldextract==5.0.1
tornado==6.3.3
traitlets==5.11.2
Twisted==22.10.0
types-protobuf==4.24.0.2
typing_extensions==4.8.0
urllib3==1.26.17
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.8
websocket-client==1.6.3
whitenoise==6.6.0
xmltodict==0.13.*
zope.interface==6.1
